feat: transpile ABI json to d.ts file
Currently, the abi json is inserted manually into the typescript file with const assertion.
The objective was to import it as const. However, this turns out to be impossible in typescript at the moment.
We introduce the script test/extract_abi.sh as an intermediary step that generates a .d.ts file from the .json abi.

See Issue #4 for more information.